+++ common/src/main/java/com/samsthenerd/hexgloop/casting/gloopifact/OpWriteGloopifact.java
@@ -1,24 +1,22 @@
 package com.samsthenerd.hexgloop.casting.gloopifact;
 
 
+import java.util.ArrayList;
 import java.util.List;
 
 import com.samsthenerd.hexgloop.casting.MishapThrowerWrapper;
-import com.samsthenerd.hexgloop.casting.truenameclassaction.HexalWispWrapper;
-import com.samsthenerd.hexgloop.casting.truenameclassaction.MishapChloeIsGonnaFindSoManyWaysToBreakThisHuh;
 import com.samsthenerd.hexgloop.items.ItemGloopifact;
 
 import at.petrak.hexcasting.api.spell.ConstMediaAction;
 import at.petrak.hexcasting.api.spell.OperationResult;
 import at.petrak.hexcasting.api.spell.casting.CastingContext;
-import at.petrak.hexcasting.api.spell.casting.CastingContext.CastSource;
 import at.petrak.hexcasting.api.spell.casting.eval.SpellContinuation;
 import at.petrak.hexcasting.api.spell.iota.BooleanIota;
 import at.petrak.hexcasting.api.spell.iota.Iota;
 import at.petrak.hexcasting.api.spell.mishaps.MishapBadOffhandItem;
-import dev.architectury.platform.Platform;
 import net.minecraft.item.ItemStack;
 import net.minecraft.text.Text;
+import net.minecraft.util.Pair;
 
 public class OpWriteGloopifact implements ConstMediaAction {
     public static List<String> expectedSources = List.of("gloopifact");
@@ -29,7 +27,7 @@ public class OpWriteGloopifact implements ConstMediaAction {
     }
 
     @Override
-    public int getArgc(){ return 0;}
+    public int getArgc(){ return 1;}
 
     @Override
     public int getMediaCost(){
@@ -52,36 +50,20 @@ public class OpWriteGloopifact implements ConstMediaAction {
 
     @Override
     public List<Iota> execute(List<? extends Iota> args, CastingContext context){
-        ItemStack castHandStack = context.getCaster().getStackInHand(context.getCastingHand());
-        if(castHandStack.getItem() instanceof ItemGloopifact gloopifactItem){
-            // make sure we're actually casting from the gloopifact
-            if(context.getSource() == CastSource.PACKAGED_HEX){
-                // verify that it's not a wisp and we should be fine ?
-                boolean isWisp = false;
-                if(Platform.isModLoaded("hexal")){
-                    isWisp = HexalWispWrapper.isWisp(context);
-                }
-                if(!isWisp){
-                    // good to go
-                    Iota iota = args.get(0);
+        Pair<ItemStack, ItemGloopifact> gloopifactLore = GloopifactUtils.assertGloopcasting(context);
+        ItemStack castHandStack = gloopifactLore.getLeft();
+        ItemGloopifact gloopifactItem = gloopifactLore.getRight();
 
-                    boolean canWrite = gloopifactItem.canWrite(castHandStack, iota);
-                    if(simulate) return List.of(new BooleanIota(canWrite));
-                    if(!canWrite){
-                        MishapThrowerWrapper.throwMishap(MishapBadOffhandItem.of(castHandStack, context.getCastingHand(), "iota.write", new Object[0]));
-                        return List.of();
-                    }
-                    gloopifactItem.writeDatum(castHandStack, iota);
-                }
-            }
-            // mishap wrong source
-            if(simulate) return List.of(new BooleanIota(false));
-            MishapThrowerWrapper.throwMishap(new MishapChloeIsGonnaFindSoManyWaysToBreakThisHuh(expectedSources));
-        } else {
-            if(simulate) return List.of(new BooleanIota(false));
-            MishapThrowerWrapper.throwMishap(MishapBadOffhandItem.of(castHandStack, context.getCastingHand(), "gloopifact"));
+        Iota iota = args.get(0);
+
+        boolean canWrite = gloopifactItem.canWrite(castHandStack, iota);
+        if(simulate) return List.of(new BooleanIota(canWrite));
+        if(!canWrite){
+            MishapThrowerWrapper.throwMishap(MishapBadOffhandItem.of(castHandStack, context.getCastingHand(), "iota.write", new Object[0]));
+            return List.of();
         }
-        return List.of();
+        gloopifactItem.writeDatum(castHandStack, iota);
+        return new ArrayList<>();
     }
 
     @Override