+++ forge/src/main/java/com/samsthenerd/hexgloop/forge/HexGloopForge.java
@@ -9,9 +9,11 @@ import com.samsthenerd.hexgloop.misc.TrinketyImplFake;
 import dev.architectury.platform.Platform;
 import dev.architectury.platform.forge.EventBuses;
 import net.minecraft.item.ItemStack;
+import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
@@ -24,6 +26,9 @@ public class HexGloopForge {
         EventBuses.registerModEventBus(HexGloop.MOD_ID, modBus);
         modBus.addListener(this::onClientSetup);
         modBus.addListener(EventPriority.NORMAL, BundleResourcePackForge::setupBuiltInResourcePack);
+
+        // yoink from hex
+        DistExecutor.unsafeRunWhenOn(Dist.CLIENT, () -> () -> modBus.register(HexGloopForgeClient.class));
 
         // setup curios
         if(Platform.isModLoaded("curios")){